refactor: 피드백반영 - 적절한 패키지로 이동
